fix(security): address all audit findings from security review
- What: Fixed Critical C-1 — truncate_str now uses char_indices() to avoid
  panics on multi-byte UTF-8 boundaries in process command names and GPU names.
  Both copies in overview.rs and processes.rs corrected.
- What: Fixed Critical C-2 — kill_process now uses i32::try_from(pid) to prevent
  negative PID (process group kill), and re-verifies /proc/PID/cmdline existence
  before sending SIGTERM to mitigate PID recycling TOCTOU.
- What: Fixed High H-1 — Replaced panic!("NVML re-init failed") with graceful
  error handling that disables GPU process monitoring instead of crashing.
- What: Fixed High H-2 — CpuTimeSample::total() and active() now use
  saturating_add/saturating_sub to prevent integer overflow in release mode.
- What: Fixed High H-3 — libc::sysconf return value now checked for -1 error,
  falling back to standard Linux default of 100 clock ticks.
- What: Fixed High H-4 — Network interface names validated against path traversal
  (reject names containing '/' or '\0').
- What: Fixed Medium M-1 — Collector errors now logged via tracing::warn instead
  of silently swallowed, making monitoring failures visible.
- What: Fixed Medium M-3 — GradientGauge label width uses unicode_width::UnicodeWidthStr
  instead of byte length for correct multi-byte character layout.
- What: Fixed Medium M-8 — Event channel changed from unbounded to bounded(256)
  to prevent memory exhaustion under input flood.
- What: Fixed Medium M-10 — /proc/PID/stat parsing now bounds-checks pos+2
  before slicing to prevent panic on malformed data.
- What: Fixed Medium M-11 — Disk sector multiplication cast to f64 before multiply
  to prevent u64 overflow with corrupted /proc/diskstats data.
- What: Fixed Low L-2 — Config deserialization now warns on parse failure instead
  of silently falling back to defaults.
- What: Fixed Low L-4 — Added AppConfig::sanitize() to clamp history_length
  (10..10000), redline_threshold (0..100), and interval (0.1..10.0), rejecting
  NaN/Infinity values to prevent DoS via malicious config file.
- Why: Security audit identified 2 Critical, 5 High, 11 Medium, 9 Low issues.

Tests: cargo test --workspace --all-features (5 passed, 0 failed).

+++ src/collectors/disk.rs
@@ -118,8 +118,10 @@ impl Collector for DiskCollector {
 
                     stats.push(DiskStats {
                         device_name: name.clone(),
-                        read_bytes_per_sec: (read_sectors_delta * SECTOR_SIZE) as f64 / elapsed,
-                        write_bytes_per_sec: (write_sectors_delta * SECTOR_SIZE) as f64 / elapsed,
+                        read_bytes_per_sec: read_sectors_delta as f64 * SECTOR_SIZE as f64
+                            / elapsed,
+                        write_bytes_per_sec: write_sectors_delta as f64 * SECTOR_SIZE as f64
+                            / elapsed,
                         read_iops: read_ios_delta as f64 / elapsed,
                         write_iops: write_ios_delta as f64 / elapsed,
                         await_read_ms: await_read,
